+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.11.1]
+### Changed
+- update README
+
 ## [0.11.0]
 ### Changed
 - refactor of `sharing-service::shareWithGroup` to call `changeGroupSharing`